+++ docs/changelog.rst
@@ -4,6 +4,7 @@ Changelog
 2.0.0 (Unreleased)
 ------------------
 
+- #38 Fix referred samples do not transition to "received at reference"
 - #37 Allow to set the default sorting order for samples in outbound shipments
 - #36 Enable configuration for notifying about retested and hidden analyses
 - #32 Fix build test after adding default contact